[PLE] fix delete tips with likes
